fix: 동물의 좋아요 수를 DB에 저장하는 방식에서 redis를 사용하는 방식으로 변경

+++ src/main/java/com/hong/ForPaw/controller/DTO/AnimalResponse.java
@@ -18,7 +18,7 @@ public class AnimalResponse {
                             String specialMark,
                             String region,
                             Long inquiryNum,
-                            Integer likeNum,
+                            Long likeNum,
                             Boolean isLike,
                             String profileURL){};
