+++ src/main/java/com/hong/ForPaw/controller/DTO/ShelterResponse.java
@@ -17,7 +17,7 @@ public class ShelterResponse {
                             String specialMark,
                             String region,
                             Long inquiryNum,
-                            Integer likeNum,
+                            Long likeNum,
                             Boolean isLike,
                             String profileURL
     ){};
